+++ packages/f-registration/CHANGELOG.md
@@ -4,9 +4,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.2.0
+------------------------------
+*April 17, 2020*
+
+### Added
+- `f-registration` base HTML structure and styling has been added.
+
+### Changed
+- Components separated out into separate `.vue` files.
+
+
 v0.1.0
 ------------------------------
 *April 7, 2020*
 
 ### Added
-- `f-registration` component structure and config
+- `f-registration` component structure and config.